Moved resampling routines into image module
