+++ sunpy/image/__init__.py
@@ -1,0 +1,3 @@
+from . resample import *
+
+__all__ = ["resample"]
